# config/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ServerCheckResponse": {
				"properties": {
					"version": {
						"type": "string"
					},
					"env": {
						"type": "string"
					},
					"requestId": {
						"type": "string"
					}
				},
				"required": [
					"version",
					"env",
					"requestId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Attachment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"originalname": {
						"type": "string"
					},
					"filename": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"path": {
						"type": "string"
					},
					"meetingId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"originalname",
					"filename",
					"type",
					"size",
					"path",
					"meetingId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Meeting": {
				"properties": {
					"id": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"startDateTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"endDateTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"description",
					"notes",
					"startDateTime",
					"endDateTime",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MeetingsResponse": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/Meeting"
						},
						"type": "array"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MeetingCreateParams": {
				"properties": {
					"description": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"startDateTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"endDateTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"description",
					"notes",
					"startDateTime",
					"endDateTime"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "simple-server",
		"version": "1.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/check": {
			"get": {
				"operationId": "GetAppCheck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServerCheckResponse"
								}
							}
						}
					}
				},
				"tags": [
					"App"
				],
				"security": [],
				"parameters": []
			}
		},
		"/meetings/{meetingId}/attachments": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Meeting"
										},
										{}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Meetings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "meetingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/meetings": {
			"get": {
				"operationId": "GetAllMeetings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeetingsResponse"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all the meetings.\nSupply the unique meeting ID and receive corresponding meeting details.",
				"tags": [
					"Meetings"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMeeting",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"endDateTime": {
											"type": "string",
											"format": "date-time"
										},
										"startDateTime": {
											"type": "string",
											"format": "date-time"
										},
										"notes": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"endDateTime",
										"startDateTime",
										"notes",
										"description",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Creates a new meeting entry in the database.\nSupply the relevant details to create the meeting details.",
				"tags": [
					"Meetings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Details with which the meeting is created",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MeetingCreateParams",
								"description": "Details with which the meeting is created"
							}
						}
					}
				}
			}
		},
		"/meetings/{meetingId}": {
			"get": {
				"operationId": "GetMeeting",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Meeting"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing meeting.\nSupply the unique meeting ID and receive corresponding meeting details.",
				"tags": [
					"Meetings"
				],
				"security": [],
				"parameters": [
					{
						"description": "The meetings identifier",
						"in": "path",
						"name": "meetingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteMeeting",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"description": "Removes the meeting from the data store.\nSupply the unique meeting ID and remove corresponding meeting details.",
				"tags": [
					"Meetings"
				],
				"security": [],
				"parameters": [
					{
						"description": "The meetings identifier",
						"in": "path",
						"name": "meetingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}